requirements.txt
Flask~=3.0.2
matplotlib~=3.8.3
pandas~=2.2.1
googletrans~=4.0.0rc1
selenium~=4.19.0
mysql-connector-python~=8.3.0
googletrans
openpyxl==3.1.2
webdriver-manager==4.0.1
pillow==10.3.0

